LayoutTests/imported/w3c/IndexedDB-private-browsing/idbindex_getKey6-expected.txt


PASS IDBIndex.getKey() - throw InvalidStateError when the index is deleted
